requirements.txt
asgiref==3.4.1
backcall==0.2.0
decorator==5.1.0
Django==3.2.9
django-debug-toolbar==3.2.2
django-extensions==3.1.5
django-filter==21.1
django-phonenumber-field==6.0.0
Faker==9.8.2
flake8==4.0.1
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-polyfill==1.0.2
ipython==7.29.0
jedi==0.18.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.2
pep8-naming==0.12.1
pexpect==4.8.0
phonenumberslite==8.12.37
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.3
q==2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
traitlets==5.1.1
wcwidth==0.2.5
